Clarify installation instructions
With some of the packaging issues fixed the installation instructions can be simplified.

In particular, `pyproject.toml` and fixed `setup.py` mean that a simple `pip install hdbscan` is now enough, assuming the user has a relatively recent `pip`. Note that the default pip on many versions of Python is pip 8, which is too old. Pip does bug to be updated, but many users do not do that. Hence there is a note that if there is an issue to first upgrade pip then try again.

Another feature of pip is that it can pull directly from GitHub. This can be useful if a new feature is merged but not yet pushed to PyPI. The one-liner pip install is more convenient than manual download and installation.