fix issue 370
It seems like we want to disallow root from being present in the eom_clusters list. I am not sure if this adversely affects the underlying algorithm, but I poured over the code and it appears that the root cannot have a lambda value? Therefore it could not be considered in the epsilon search function.