Integrate-graph-objects into hdbscan_.py
Create a new option for metrics calles "graph", which takes graph objects as a csr adjacency matrix and runs the HDBSCAN function on the csgraph min_span_tree of the given graph.

Example plot file (plot_hdbscan_graph.py) shows the working of the function and displays the example plots of the graphs.