Adds cythonize() call to extensions in setup.py()
I've been having trouble build hdbscan on MacOS with normal
python (e.g. outside of conda). It turns out that nothing was
instructing the build system to actually compile the cython files.

After reading the cython documentation, it seems like the easiest
fix is to wrap the extensions list in cythonize(). Locally this
built fine.